# 23% faster Nil#try