@schneems 2019-02-21
Wanted to link that this is a continued effort of two other PRs

- https://github.com/rails/rails/pull/33747
- https://github.com/rails/rails/pull/33752