@schneems 2021-07-13
> This PR speeds up Nil#try by avoiding an allocation when only one argument is passed:

FWIW the array is allocated no matter what. Also apparently this is faster in Ruby 2.5 but slower in 2.7 and 3.0 https://twitter.com/jeremyevans0/status/1414731600693108737